+++ lib/router/src/models/ny_page_transition_settings.dart
@@ -1,3 +1,4 @@
+import 'package:flutter/cupertino.dart' show CupertinoPageTransitionsBuilder;
 import 'package:flutter/material.dart';
 
 /// This class contains all page transition settings.