+++ lib/router/src/page_transition/src/page_transition.dart
@@ -1,4 +1,5 @@
 import 'dart:io';
+import 'package:flutter/cupertino.dart' show CupertinoPageTransitionsBuilder;
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'enum.dart';